Updated logging format
Altered the logging format for easier readability and
parsability.  The code was overcomplicated previously
and now outputs the time in one column, the log level in
another column, and the actual message in the final column.
These are space separated, with more or less a finite number
of spaces for each column.