pid_file() bug and logging macros
A bug in pid_file prevented the file from being removed
when the application exited.  While this was added to
terminate(), we may need to later add it to 'atexit'
for better clean up.

Also added a few macros for logging such as debug2(),
debug(), and critical().  The configuration is dumped
to debug2() immediately after starting.  Debug2 should
be considered more verbose than debug().