Addition of Stanzas and Mode on command line.
From the command line, you now have the option of using
-m for ``mode.''  Valid modes are currently master or
distro.

A new "stanza" support was added.  If you specify -s
on the command line, and a name, it will scan the config
for that particular stanza and use options located there.

For example:
 stagemediad -s mystanza

In the configuration:
 <will_not_be_used>
 option1=value1
 option2=value2
 </will_not_be_used>

 <mystanza>
 option1=value1
 option2=value2
 </mystanza>

In the above, options from the 'mystanza' will be used, while
options from ``will_not_be_used'' will be ignored.