Fixup Dead Sockets
In some cases, sockets would connect and the server would just
sit there and spin around, encoding into mp3.  However, it never
got delivered and was consuming a file descriptor for no good reason.
The "fix" (kludge) was to add non-blocking I/O to the socket temporarily
and, if it error'd out, abort further processing.